- refactor networking class to its own package - slight update to model
